+++ test/test_type_names.dart
@@ -4,7 +4,7 @@ void main() {
   var t = new BinaryTypes();
   var dataModel = t["int"].dataModel;
   var types = <BinaryType>[];
-  var type = new FunctionType("myfunc", t["int"], [(t["void*"])], dataModel);
+  var type = new FunctionType("myfunc", t["int"], [(t["void*"])], false, dataModel);
   types.add(type);
   types.add(type.ptr());
   types.add(type.ptr().ptr());
@@ -17,7 +17,7 @@ void main() {
   types.add(type.array(1).array(2).ptr());
   types.add(type.array(1).array(2).ptr().ptr());
   types.add(type.array(1).ptr().array(1));
-  types.add(new EnumType("Color", dataModel)..addEnumerators(["RED", "GREEN", "RED"]));
+  types.add(new EnumType("Color", dataModel)..addEnumerators(["RED", "GREEN", "BLUE"]));
   print("=====================");
   print("Types:");
   for (var type in types) {